separate KV logic
